PMM-6852 move notifiers to a separate files
